requirements.txt
chalice
boto3
botocore
websocket-client
pytest
requests